Implement the new core
