+++ src/component.js
@@ -1,4 +1,3 @@
-import { Container } from './core.js';
 
 function toCamelCase(val = '') {
   return val.indexOf('-') === -1 ? val.toLowerCase() : val.toLowerCase().split('-').reduce((out, part) => {
@@ -6,81 +5,85 @@ function toCamelCase(val = '') {
   },'') 
 }
 
-function component(renderer, BaseElement = HTMLElement, {useShadowDOM = true, shadowRootInit = {}} = {}) {
-  class Element extends BaseElement {
-    static get observedAttributes() {
-      return renderer.observedAttributes || [];
-    }
-
-    constructor() {
-      super();
-      if (useShadowDOM === false) {
-        this._container = new Container(renderer, this);
-      } else {
-        this.attachShadow({ mode: "open", ...shadowRootInit});
-        this._container = new Container(renderer, this.shadowRoot, this);        
+function makeComponent(Container) {
+  function component(renderer, BaseElement = HTMLElement, {useShadowDOM = true, shadowRootInit = {}} = {}) {
+    class Element extends BaseElement {
+      static get observedAttributes() {
+        return renderer.observedAttributes || [];
       }
-    }
 
-    connectedCallback() {
-      this._container.update();
-    }
+      constructor() {
+        super();
+        if (useShadowDOM === false) {
+          this._container = new Container(renderer, this);
+        } else {
+          this.attachShadow({ mode: "open", ...shadowRootInit});
+          this._container = new Container(renderer, this.shadowRoot, this);        
+        }
+      }
 
-    disconnectedCallback() {
-      this._container.teardown();
-    }
-
-    attributeChangedCallback(name, _, newValue) {
-      let val = newValue === '' ? true : newValue;
-      Reflect.set(this, toCamelCase(name), val);
-    }
-  };
-
-  function reflectiveProp(initialValue) {
-    let value = initialValue;
-    return Object.freeze({
-      enumerable: true,
-      configurable: true,
-      get() {
-        return value;
-      },
-      set(newValue) {
-        value = newValue;
+      connectedCallback() {
         this._container.update();
       }
-    })
+
+      disconnectedCallback() {
+        this._container.teardown();
+      }
+
+      attributeChangedCallback(name, _, newValue) {
+        let val = newValue === '' ? true : newValue;
+        Reflect.set(this, toCamelCase(name), val);
+      }
+    };
+
+    function reflectiveProp(initialValue) {
+      let value = initialValue;
+      return Object.freeze({
+        enumerable: true,
+        configurable: true,
+        get() {
+          return value;
+        },
+        set(newValue) {
+          value = newValue;
+          this._container.update();
+        }
+      })
+    }
+
+    const proto = new Proxy(BaseElement.prototype, {
+      set(target, key, value, receiver) {
+        if(key in target) {
+          Reflect.set(target, key, value);
+        }
+        let desc;
+        if(typeof key === 'symbol' || key[0] === '_') {
+          desc = {
+            enumerable: true,
+            configurable: true,
+            writable: true,
+            value
+          }; 
+        } else {
+          desc = reflectiveProp(value);
+        }
+        Object.defineProperty(receiver, key, desc);
+
+        if(desc.set) {
+          desc.set.call(receiver, value);
+        }
+
+        return true;
+      }
+    });
+
+    Object.setPrototypeOf(Element.prototype, proto);
+
+
+    return Element;
   }
 
-  const proto = new Proxy(BaseElement.prototype, {
-    set(target, key, value, receiver) {
-      if(key in target) {
-        Reflect.set(target, key, value);
-      }
-      let desc;
-      if(typeof key === 'symbol' || key[0] === '_') {
-        desc = {
-          enumerable: true,
-          configurable: true,
-          writable: true,
-          value
-        }; 
-      } else {
-        desc = reflectiveProp(value);
-      }
-      Object.defineProperty(receiver, key, desc);
-
-      if(desc.set) {
-        desc.set.call(receiver, value);
-      }
-
-      return true;
-    }
-  });
-
-  Object.setPrototypeOf(Element.prototype, proto);
-
-
-  return Element;
+  return component;
 }
 
-export { component };
+export { makeComponent };
