+++ src/haunted.js
@@ -1,12 +1,3 @@
-export { component } from './component.js';
-export { html, render } from './core.js';
-export { useCallback } from "./use-callback.js";
-export { useEffect } from './use-effect.js';
-export { useState } from './use-state.js';
-export { useReducer } from './use-reducer.js';
-export { useMemo } from './use-memo.js';
-export { useRef } from './use-ref.js';
-export { withHooks, virtual } from './virtual.js';
-export { useContext } from './use-context.js';
-export { createContext } from './create-context.js';
-export { hook, Hook } from './hook.js';
+export { html, render, component, createContext, virtual } from './lit-haunted.js';
+export * from './core.js';
+export { default } from './core.js';
