+++ src/lit-haunted.js
@@ -1,0 +1,19 @@
+import { html, render } from 'lit-html';
+import haunted from './core.js';
+import { makeVirtual } from './virtual.js';
+
+const { Container, component, createContext } = haunted({
+  render(what, where) {
+    render(what, where);
+  }
+});
+
+const virtual = makeVirtual(Container);
+
+export {
+  component,
+  createContext,
+  virtual,
+  html,
+  render
+};
