+++ src/virtual.js
@@ -1,48 +1,49 @@
-import { Container } from './core.js';
-import { directive } from './lit.js';
-
-const partToContainer = new WeakMap();
-const containerToPart = new WeakMap();
-
-class DirectiveContainer extends Container {
-  constructor(renderer, part) {
-    super(renderer, part);
-    this.virtual = true;
-  }
-
-  commit(result) {
-    this.host.setValue(result);
-    this.host.commit();
-  }
-
-  teardown() {
-    super.teardown();
-    let part = containerToPart.get(this);
-    partToContainer.delete(part);
-  }
-}
-
-
-
-function withHooks(renderer) {
-  function factory(...args) {
-    return part => {
-      let cont = partToContainer.get(part);
-      if(!cont) {
-        cont = new DirectiveContainer(renderer, part);
-        partToContainer.set(part, cont);
-        containerToPart.set(cont, part);
-        teardownOnRemove(cont, part);
-      }
-      cont.args = args;
-      cont.update();
-    };
-  }
-
-  return directive(factory);
-}
+import { directive } from 'lit-html';
 
 const includes = Array.prototype.includes;
+
+function makeVirtual(Container) {
+  const partToContainer = new WeakMap();
+  const containerToPart = new WeakMap();
+  
+  class DirectiveContainer extends Container {
+    constructor(renderer, part) {
+      super(renderer, part);
+      this.virtual = true;
+    }
+  
+    commit(result) {
+      this.host.setValue(result);
+      this.host.commit();
+    }
+  
+    teardown() {
+      super.teardown();
+      let part = containerToPart.get(this);
+      partToContainer.delete(part);
+    }
+  }
+  
+  function virtual(renderer) {
+    function factory(...args) {
+      return part => {
+        let cont = partToContainer.get(part);
+        if(!cont) {
+          cont = new DirectiveContainer(renderer, part);
+          partToContainer.set(part, cont);
+          containerToPart.set(cont, part);
+          teardownOnRemove(cont, part);
+        }
+        cont.args = args;
+        cont.update();
+      };
+    }
+  
+    return directive(factory);
+  }
+
+  return virtual;
+}
 
 function teardownOnRemove(cont, part, node = part.startNode) {
   let frag = node.parentNode;
@@ -67,4 +68,4 @@ function teardownOnRemove(cont, part, node = part.startNode) {
   mo.observe(frag, { childList: true });
 }
 
-export { withHooks, withHooks as virtual }
+export { makeVirtual };
